Add validation if no post is found when trying to destroy

+++ app/controllers/posts_controller.rb
@@ -9,13 +9,17 @@ class PostsController < ApplicationController
     return render json: new_post, status: :ok if new_post.save
     
     render json: {errors: new_post.errors.full_messages}, status: :unprocessable_entity
-  
   end
 
   def destroy
-    post_to_destroy = Post.find(params[:id])
-    post_to_destroy.destroy
-    render json: post_to_destroy, status: :ok
+
+    post_to_destroy = Post.find_by(id: params[:id])
+    if post_to_destroy
+      post_to_destroy.destroy
+      render json: post_to_destroy, status: :ok
+    else
+      render json: {error: 'Could not find the post id'}, status: :unprocessable_entity
+    end
   end
 
   private
